requirements.txt
flask~=2.0.2
flask-restx
Werkzeug~=2.0.2
requests==2.26.0
Pillow~=8.4.0